Fix count field not being populated in HaloBox via CIC interpolation
Agent-Logs-Url: https://github.com/21cmfast/21cmFAST/sessions/ec97c833-ba5c-4c87-b9f5-fca0e97b3f45

+++ src/py21cmfast/src/_outputstructs_wrapper.h
@@ -49,7 +49,7 @@ typedef struct HaloBox {
     float *halo_mass;
     float *halo_stars;
     float *halo_stars_mini;
-    int *count;
+    float *count;
 
     // For IonisationBox.c and SpinTemperatureBox.c
     float *n_ion;     // weighted by F_ESC*PopN_ion
